convert flask comms class to function
